Move createWretchInstance to package
