+++ apps/frontend/lib/utils/api.ts
@@ -1,10 +1,6 @@
+import { createWretchInstance } from '@repo/api';
 import hash from 'object-hash';
 import { toast } from 'sonner';
-import wretch, { Wretch, WretchResponseChain } from 'wretch';
-import AbortAddon, { AbortResolver, AbortWretch } from 'wretch/addons/abort';
-import QueryStringAddon, {
-  QueryStringAddon as QueryStringAddonType,
-} from 'wretch/addons/queryString';
 import { WretchError } from 'wretch/resolver';
 
 import { NextApiRequestUnknown } from '@pages/api/types';
@@ -27,16 +23,6 @@ type CacheSettings = {
   useStartEndDates?: boolean;
   deDupeKey?: string;
 };
-
-type WretchInstance = QueryStringAddonType &
-  AbortWretch &
-  Wretch<AbortWretch & QueryStringAddonType, AbortResolver, undefined>;
-type WretchResponse = AbortResolver &
-  WretchResponseChain<
-    AbortWretch & QueryStringAddonType,
-    AbortResolver,
-    undefined
-  >;
 
 // Helper functions
 function setWithExpiry<T>(
@@ -217,48 +203,6 @@ function handleCacheSet<Query, Response>({
     storageType,
     cache?.useStartEndDates ? start_end_dates : undefined,
   );
-}
-
-function createWretchInstance<Query, Body>({
-  url,
-  method,
-  query,
-  body,
-  controller,
-  bearerToken,
-}: {
-  url: string;
-  method: 'GET' | 'POST' | 'DELETE' | string;
-  controller: AbortController;
-  query?: Query;
-  body?: Body;
-  bearerToken?: string;
-}) {
-  let wretchInstance: WretchInstance | WretchResponse = wretch()
-    .url(url)
-    .addon(AbortAddon())
-    .addon(QueryStringAddon)
-    .signal(controller)
-    .query(query || {});
-
-  if (bearerToken) {
-    wretchInstance = wretchInstance.auth(`Bearer ${bearerToken}`);
-  }
-
-  switch (method) {
-    case 'GET':
-      wretchInstance = wretchInstance.get();
-      break;
-    case 'POST':
-      wretchInstance = wretchInstance.json(body || {}).post();
-      break;
-    case 'DELETE':
-      wretchInstance = wretchInstance.json(body || {}).delete();
-      break;
-    default:
-      throw new Error('Invalid method');
-  }
-  return wretchInstance;
 }
 
 function processCachedResponse<Query, Response>(
@@ -469,4 +413,4 @@ function getQueryParam(query: NextApiRequestUnknown['query'], param: string) {
   return Array.isArray(value) ? undefined : value;
 }
 
-export { regularFetch, getQueryParam, createWretchInstance };
+export { regularFetch, getQueryParam };
