+++ apps/frontend/lib/utils/strava/authentication.ts
@@ -1,6 +1,5 @@
+import { createWretchInstance } from '@repo/api';
 import wretch from 'wretch';
-
-import { createWretchInstance } from '@utils/api';
 
 import { baseUrl } from './config';
 import { StravaAuthResponse } from './types';
