+++ apps/frontend/package.json
@@ -25,6 +25,7 @@
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
+    "@repo/api": "*",
     "@repo/cosmosdb": "*",
     "@repo/jwt": "*",
     "@repo/schemas": "*",
