+++ packages/api/index.ts
@@ -1,0 +1,3 @@
+import { createWretchInstance } from './src/wretch';
+
+export { createWretchInstance };
