+++ packages/api/readme.md
@@ -1,0 +1,18 @@
+# API
+
+This package contains API utilities for the repository.
+
+## Usage
+
+```typescript
+import { createWretchInstance } from '@repo/api';
+
+const wretchInstance = createWretchInstance({
+    url: `https://api.example.com`,
+    method: 'GET',
+    query: {
+        key: 'value',
+    },
+    controller: new AbortController(),
+});
+```
